move TokenPruningConfig to base _types level
