+++ specification/_types/TokenPruningConfig.ts
